Extract Updater login into separate class.
Move various enums and interfaces to organized subdirectories under Models and Interfaces namespaces. Extract the update logic from the Program class into a new, dedicated Updater class to improve code modularity and maintainability. Also, update affected import statements accordingly.
