+++ src/DependencyUpdated.Core/Models/DependencyDetails.cs
@@ -1,3 +1,3 @@
-﻿namespace DependencyUpdated.Core;
+﻿namespace DependencyUpdated.Core.Models;
 
 public record DependencyDetails(string Name, Version Version);
