+++ src/DependencyUpdated.Core/Models/Enums/ProjectType.cs
@@ -1,0 +1,6 @@
+namespace DependencyUpdated.Core.Models.Enums;
+
+public enum ProjectType
+{
+    DotNet = 1,
+}
