+++ src/DependencyUpdated.Core/Models/UpdateResult.cs
@@ -1,3 +1,3 @@
-namespace DependencyUpdated.Core;
+namespace DependencyUpdated.Core.Models;
 
 public record UpdateResult(string PackageName, string OldVersion, string NewVersion);
