--- src/DependencyUpdated.Core/ProjectType.cs
@@ -1,6 +1,0 @@
-namespace DependencyUpdated.Core;
-
-public enum ProjectType
-{
-    DotNet = 1,
-}
