+++ src/DependencyUpdated/Program.cs
@@ -8,7 +8,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Serilog;
 using System.ComponentModel.DataAnnotations;
-using System.IO.Enumeration;
 using ILogger = Serilog.ILogger;
 
 namespace DependencyUpdated;
@@ -37,68 +36,9 @@ public static class Program
             throw new OptionsValidationException(nameof(UpdaterConfig), typeof(UpdaterConfig),
                 validationResult.Where(x => !string.IsNullOrEmpty(x.ErrorMessage)).Select(x => x.ErrorMessage!));
         }
-        
-        var repositoryProvider =
-            _serviceProvider.GetRequiredKeyedService<IRepositoryProvider>(config.Value.RepositoryType);
-        var repositoryPath = Environment.CurrentDirectory;
-        repositoryProvider.SwitchToDefaultBranch(repositoryPath);
 
-        foreach (var configEntry in config.Value.Projects)
-        {
-            var updater = _serviceProvider.GetRequiredKeyedService<IProjectUpdater>(configEntry.Type);
-
-            foreach (var directory in configEntry.Directories)
-            {
-                if (!Path.Exists(directory))
-                {
-                    throw new FileNotFoundException("Search path not found", directory);
-                }
-
-                var projectFiles = updater.GetAllProjectFiles(directory);
-                var allDependenciesToUpdate =
-                    await updater.ExtractAllPackagesThatNeedToBeUpdated(projectFiles, configEntry);
-
-                if (allDependenciesToUpdate.Count == 0)
-                {
-                    continue;
-                }
-
-                var uniqueListOfDependencies = allDependenciesToUpdate.DistinctBy(x => x.Name).ToList();
-                var projectName = ResolveProjectName(configEntry, directory);
-                foreach (var group in configEntry.Groups)
-                {
-                    var matchesForGroup = uniqueListOfDependencies
-                        .Where(x => FileSystemName.MatchesSimpleExpression(group, x.Name)).ToArray();
-                    if (matchesForGroup.Length == 0)
-                    {
-                        continue;
-                    }
-                    
-                    uniqueListOfDependencies.RemoveAll(x => FileSystemName.MatchesSimpleExpression(group, x.Name));
-                    repositoryProvider.SwitchToUpdateBranch(repositoryPath, projectName, group);
-
-                    var allUpdates = updater.HandleProjectUpdate(projectFiles, matchesForGroup);
-                    if (allUpdates.Count == 0)
-                    {
-                        continue;
-                    }
-
-                    repositoryProvider.CommitChanges(repositoryPath, projectName, group);
-                    await repositoryProvider.SubmitPullRequest(allUpdates.DistinctBy(x => x.PackageName).ToArray(), projectName, group);
-                    repositoryProvider.SwitchToDefaultBranch(repositoryPath);
-                }
-            }
-        }
-    }
-
-    private static string ResolveProjectName(Project project, string directory)
-    {
-        if (!project.EachDirectoryAsSeparate)
-        {
-            return project.Name;
-        }
-
-        return Path.GetFileName(directory);
+        var updater = ActivatorUtilities.CreateInstance<Updater>(_serviceProvider);
+        await updater.DoUpdate();
     }
 
     private static void Configure(Options appOptions)
@@ -122,7 +62,7 @@ public static class Program
             .AddSingleton(TimeProvider.System)
             .AddMemoryCache()
             .RegisterDotNetServices()
-            .RegisterGithub()
+            .RegisterAzureDevOps()
             .AddOptions<UpdaterConfig>().Bind(_configuration.GetSection("UpdaterConfig"));
 
         _serviceProvider = services.BuildServiceProvider();
